Extract test from stories_index_spec to tagged_articles_spec
